+++ yash-syntax/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to `yash-syntax` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - ????-??-??
+
+### Removed
+
+- `source::Location::advance`
+
 ## [0.3.0] - 2022-02-06
 
 This version simplifies type definitions for the abstract syntax tree (AST);